backend/requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.2.0
django-environ==0.8.1
httpie==3.1.0
drf-yasg==1.20.0